Move affix file parsing into submod
