+++ speller/src/hunspell/parse_aff.rs
