+++ speller/src/lib.rs
@@ -2,7 +2,6 @@ use anyhow::Result;
 use std::path::Path;
 
 mod hunspell;
-mod hunspell_aff;
 
 pub use crate::hunspell::SpellerHunspellDict;
